+++ src/app/main/apps/modelos/modelo/store/effects/componentes-digitais.effects.ts
@@ -102,11 +102,23 @@ export class ComponenteDigitalEffect {
             1,
             0,
             '{}',
-            '[]')),
+            JSON.stringify([
+                'processoOrigem',
+                'tarefaOrigem'
+            ])
+        )),
         switchMap(response => [
-            new AddData<Documento>({data: response['entities'], schema: documentoSchema}),
+            new AddData<Documento>({
+                data: response['entities'],
+                schema: documentoSchema,
+                populate: [
+                    'processoOrigem',
+                    'tarefaOrigem'
+                ]
+            }),
             new ComponenteDigitalActions.GetDocumentoSuccess({
                 documentoId: response['entities'][0].id,
+                documento: response['entities'][0],
                 componenteDigitalId: this.componenteDigitalId,
                 routeTarefa: this.routeAtividadeTarefa,
                 routeDocumento: this.routeAtividadeDocumento
@@ -120,37 +132,33 @@ export class ComponenteDigitalEffect {
     getDocumentoSuccess: any = createEffect(() => this._actions.pipe(
         ofType<ComponenteDigitalActions.GetDocumentoSuccess>(ComponenteDigitalActions.GET_DOCUMENTO_SUCCESS),
         tap((action) => {
-            const primary = 'componente-digital/' + action.payload.componenteDigitalId;
-            const sidebar = 'editar/' + action.payload.routeDocumento;
+            let stepHandle = this.routerState.params['stepHandle'];
+            let primary: string;
+            primary = 'componente-digital/';
+            const componenteDigitalId = action.payload.componenteDigitalId;
 
-            if (this.routerState.url.indexOf('processo') !== -1) {
-                this._router.navigate([
-                        this.routerState.url.split('processo/')[0] + 'processo/' + this.routerState.params.processoHandle
-                        + '/visualizar/' + this.routerState.params.stepHandle + '/documento/' + action.payload.documentoId,
-                        {
-                            outlets: {
-                                primary: primary,
-                                sidebar: sidebar
-                            }
-                        }
-                    ],
-                    {
-                        relativeTo: this._activatedRoute.parent
-                    }).then();
-            } else {
-                this._router.navigate([
-                        this.routerState.url.replace('modelos/modelo', action.payload.routeTarefa + '/documento') + '/' + action.payload.documentoId,
-                        {
-                            outlets: {
-                                primary: primary,
-                                sidebar: sidebar
-                            }
-                        }
-                    ],
-                    {
-                        relativeTo: this._activatedRoute.parent
-                    }).then();
+            primary += componenteDigitalId;
+
+            const tarefaId = action.payload.documento.vinculacaoDocumentoPrincipal ?
+                action.payload.documento.vinculacaoDocumentoPrincipal.documento.tarefaOrigem.id :
+                action.payload.documento.tarefaOrigem.id;
+            const processoId = action.payload.documento.processoOrigem.id;
+            if (!stepHandle || processoId !== parseInt(this.routerState.params['processoHandle'], 10)) {
+                stepHandle = 'default';
             }
+            this._router.navigate([
+                    'apps/tarefas/' + this.routerState.params.generoHandle + '/' + this.routerState.params.typeHandle + '/'
+                    + this.routerState.params.targetHandle + '/tarefa/' + tarefaId + '/processo/' + processoId + '/visualizar/'
+                    + stepHandle + '/documento/' + action.payload.documento.id,
+                    {
+                        outlets: {
+                            primary: primary
+                        }
+                    }
+                ],
+                {
+                    relativeTo: this._activatedRoute.parent
+                }).then();
         })
     ), {dispatch: false});
     /**